Refactor actions to use the plugin model
Each action is now defined in a single file. The file must export a `card` object, a `handler` function and an optional `pre` function. These actions are then given the plugin's base class constructor and exposed by the plugin's `getActions` method.

Obviously this is a major change. It will need to be released in conjunction with other major changes that update how we handle actions!

Change-type: major
Signed-off-by: Graham McCulloch <graham@balena.io>